Added: Project Structure & NuGets
Added the base structure of the system by creating all of the various projects, and making a few alterations and basic code to continue working on the different classes and gradually build up the system.

Also added the BaseUser class which the Influencer and Company will inherit from to the DataAccessLayer.

Also added Nuget packages in the form of RestSharp, Microsoft SQL, Dapper.