Added: Create Influencer Functionality
Added the functionality for creating an influencer, including making the view and setting certain restrictions on what can be added to the input fields.

Also added a tiny nice-to-have user functionality feature in, which makes all domains start with hashtag, and removes their spaces, and divides this in to three seperate input fields.

Also played around with some successmessage and tempdata, and made it so when something is a success a notification will be made. The coloring are not ideal quite yet, but if you like the feature we can look at improving the design.