Changed: Visual Design
Applied a lot more visual design to the website in the form of a sign-in and sign-up page with an introduction text and a login screen on the right side.

The influencer/create page has now been properly styled as well, and follows the websites theme.

The funny little TempData based notification has been updated to have a more fitting look and match the overall theme.