Changed: SQL Scripts
Made changes to all of the SQL scripts in the form of modifying the maximum input values as the requirements of the system became more clear, this including setting influencer domains, company domains and announcement subjects to only be 20 characters long.

Influencer and Company biography fields can now maximum be 500 characters long.