Added: Tests For Announcement
Added tests for the announcement feature, the tests include being able to validate that a new announcement has been created and persisted in the database, a new announcement can be created with its subjects inserted in to the announcementsubjects table and a test to prove that when wrong information is inserted in to the subject then the whole transaction is rolled back to a previously stable state.

Also made minor changes to the InfluencerDaoTest class to make it more readable, but will have to add more comments later on to make it easier to read for the whole team.