Added: Views For WinForms
Added all of the views that will be used in the WinForms application, along with a lot of the functionality, but only on a demo-testing level. The form is not connected with the API yet, but seemingly the functionality of most of it seem to be working now at least visually.