Added: Announcement Overview (80% Complete)
Added a view where it is possible to get an overview of all of the announcements currently in the database.

This feature was not entirely completed within this sprint, and will need to maybe have some changes made, it surely need to have some tests made for it, although we can see right now that it works as intended for now.

This is only being pushed to the main branch because we have to present what we managed in sprint 1.