Added: Announcement GetOne Test
Added a test for the Announcement GetOne functionality, which is used to retrieve information regarding an announcement, the name of the company that created the announcement, along with the subjects associated with the announcement.

The test tests up against our insert data and is able to find the correct information from the database, showing that the correct title, the company's name along with the subjects can all be found and are the expected values.