Updated: Sticky-footer
Different places in the website, we used a home-footer in the top, to make the footer stay at the bottom.
When this class for made in site, i thought it was only going to be used in the frontpage. so it was called home-footer. Which was a bit unclear. So i changed it to sticky-footer - cause its being used different places.