Added: Optimistic Concurrency for Announcement
Implemented optimistic concurrency control for insertion of newly created announcements into the database.

The optimistic concurrency is handled through the usage of implementing RowVersion in to the Announcements database table,

Also changed a few lorem impsum texts and input choice options.

Changed the link destination for when you click on the logo, and also changed the routing so that you start on the announcement controller's index action as the home / landing page.