Added: Delete Functionality (Only) and Test Fixes
Solely added the deletion functionality itself, the delete button's actual functionality in the WinForms dashboard will be handled later in this iteration, as it would make more sense to have completed another user story first as this feature interacts with the table that should display the announcements.

Also fixed all of the tests, so that they pass and work as intended, after finding out the tests were not functional with the previously created test data and upon re-generating the test data and further investigation we found that the test proving concurrency issues was handled, was passing, but actually was not correctly proving that the concurrency was handled, this has now been solved.