Changed: Explicit Isolation Levels
After reading up on the isolation levels after the feedback received at the last sprint review, we have delved further in to the isolation levels, and although we already are using ReadCommitted as it is the default isolation level, we have added it explicitedly in the code.

We were advised to look in to possibly use ReadUncommitted after a hint by Karsten, but after reading through the referenced power point materials, and being unable to find what may have been referenced as a possible solution to using ReadUncommitted without introducing unncessary compromises in the system we have chosen to continue with using ReadCommitted.